LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/autofocus/focusable-area-in-top-document-expected.txt



FAIL If topDocument's focused area is not topDocument, autofocus is not processed. assert_equals: activeElement should not be changed expected Element node <iframe srcdoc="<input><script>document.querySelector('in... but got Element node <input autofocus=""></input>
